EventService/requirements/base.txt
asgiref==3.8.1
Django==4.2.11
django-environ==0.11.2
django-extensions==3.2.3
django-model-utils==4.5.0
django-storages==1.14.2
djangorestframework==3.15.1
djoser==2.2.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
structlog
typing_extensions==4.10.0
uritemplate==4.1.1

